Convert bg music to ADCPM using adpcm-qx
https://github.com/dbry/adpcm-xq
Using default GoldWave or Audacity settings cuts off some parts of the file
since ADPCM is block based. So converted wavs no longer loop properly
With block size set to 256 the difference is minimal and not audible.
(-b8 option for adpcm-hq)

Add script to convert all wav in a folder